started implementing typescript
