+++ server/package.json
@@ -5,19 +5,25 @@
   "main": "app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start:dev": "NODE_ENV=development nodemon app.js",
-    "start" : "node app.js"
+    "dev:w": "set NODE_ENV=development&&nodemon app.js",
+    "dev": "NODE_ENV=development nodemon app.js",
+    "prod": "node app.js"
   },
   "author": "Kapil",
   "license": "ISC",
   "dependencies": {
     "cors": "^2.8.5",
-    "dotenv": "^16.0.3",
-    "express": "^4.18.1",
-    "joi": "^17.6.0",
-    "jsonwebtoken": "^8.5.1",
-    "mongoose": "^6.5.3",
-    "nodemon": "^2.0.19",
-    "uuid": "^8.3.2"
+    "dotenv": "^16.4.5",
+    "express": "^4.18.3",
+    "joi": "^17.12.2",
+    "jsonwebtoken": "^9.0.2",
+    "mongoose": "^8.2.0",
+    "nodemon": "^3.1.0",
+    "uuid": "^9.0.1"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.21",
+    "@types/node": "^20.11.23",
+    "typescript": "^5.3.3"
   }
 }
